+++ src/tightdb/column_type.hpp
@@ -50,13 +50,17 @@ enum ColumnAttr {
     col_attr_None = 0,
     col_attr_Indexed = 1,
 
-    // Specifies that this column forms a unique constraint. It requires
-    // `col_attr_Indexed`.
+    /// Specifies that this column forms a unique constraint. It requires
+    /// `col_attr_Indexed`.
     col_attr_Unique = 2,
 
-    // Specifies that this column forms the primary key. It implies a non-null
-    // constraint on the column, and it requires `col_attr_Unique`.
-    col_attr_PrimaryKey = 4
+    /// Specifies that this column forms the primary key. It implies a non-null
+    /// constraint on the column, and it requires `col_attr_Unique`.
+    col_attr_PrimaryKey = 4,
+
+    /// Specifies that the links of this column are strong, not weak. Applies
+    /// only to link columns (`type_Link` and `type_LinkList`).
+    col_attr_StrongLinks = 8
 };