+++ test/util/test_path.cpp
@@ -74,9 +74,14 @@ TestPathGuard::~TestPathGuard() TIGHTDB_NOEXCEPT
 SharedGroupTestPathGuard::SharedGroupTestPathGuard(const string& path):
     TestPathGuard(path)
 {
-    File::try_remove(m_path+".lock");
+    File::try_remove(get_lock_path());
 }
 
+
+SharedGroupTestPathGuard::~SharedGroupTestPathGuard()
+{
+    File::try_remove(get_lock_path());
+}
 
 } // namespace test_util
 } // namespace tightdb